Move metta_atom and metta_space out of the common module
Move metta_atom into the text module, metta_space into the grounding
space module.
