+++ lib/src/space/grounding/mod.rs
@@ -254,6 +254,19 @@ impl CustomMatch for GroundingSpace {
 }
 
 #[cfg(test)]
+use crate::metta::text::*;
+
+#[cfg(test)]
+pub(crate) fn metta_space(text: &str) -> DynSpace {
+    let mut space = GroundingSpace::new();
+    let mut parser = SExprParser::new(text);
+    while let Some(atom) = parser.parse(&Tokenizer::new()).unwrap() {
+        space.add(atom);
+    }
+    space.into()
+}
+
+#[cfg(test)]
 mod test {
     use super::*;
     use crate::matcher::*;
